+++ Client/src/utils/expeditionDataUtils.ts
@@ -1,5 +1,11 @@
 import { ExpeditionCharacter } from '../types/Types'
 import { levelRanges } from '../constants/levelRanges'
+import { RaidGold, raidGoldTable, RaidType } from '../constants/goldRaidTable'
+
+interface AvailableRaid {
+  raidName: string
+  type: '노말' | '하드' | '싱글'
+}
 
 // 원정대 상위 6명 캐릭터의 평균 레벨 계산
 export const calculateAverageLevel = (characters?: ExpeditionCharacter[]): number => {
@@ -70,4 +76,22 @@ export const getGoldByLevelRange = (data: { name: string; level: string; gold: n
       }
     })
     .filter((item) => item.totalGold > 0)
+}
+
+// 레이드 레벨에 따른 활성화 가능한 레이드 목록을 반환하는 함수
+export const getAvailableRaidsByLevel = (level: number): AvailableRaid[] => {
+  const available: AvailableRaid[] = []
+
+  raidGoldTable.forEach((raid: RaidGold) => {
+    raid.type.forEach((raidType: RaidType) => {
+      if (level >= raidType.level) {
+        available.push({
+          raidName: raid.name,
+          type: raidType.type,
+        })
+      }
+    })
+  })
+
+  return available
 }
